Controller test coverage

+++ src/main/java/com/deliverar/pagos/domain/dtos/TransactionDto.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -19,7 +20,7 @@ public class TransactionDto {
     private UUID id;
     private OwnerDto originOwner;
     private OwnerDto destinationOwner;
-    private BigDecimal amount;
+    private BigInteger amount;
     private CurrencyType currency;
     private BigDecimal conversionRate;
     private String concept;
